+++ src/components/index.js
@@ -36,9 +36,6 @@ export { Icon }
 import Code from './atoms/code'
 export { Code }
 
-import EmptyState from './atoms/empty-state'
-export { EmptyState }
-
 import {
   Heading1,
   Heading2,
@@ -51,6 +48,9 @@ import {
 export { Heading1, Heading2, Heading3, Heading4, Heading5, Subheader, Text }
 
 /* molecules */
+import EmptyState from './molecules/empty-state'
+export { EmptyState }
+
 import Form from './molecules/form'
 export { Form }
